chore: organized files, fixes
