+++ src/main/java/com/good/food/gateways/http/ClienteController.java
@@ -14,8 +14,8 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import com.good.food.gateways.http.request.ClienteRequest;
 import com.good.food.gateways.http.response.ClienteResponse;
-import com.good.food.usecase.BuscarCliente;
-import com.good.food.usecase.CadastrarCliente;
+import com.good.food.usecase.cliente.BuscarCliente;
+import com.good.food.usecase.cliente.CadastrarCliente;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
