+++ src/main/java/com/good/food/usecase/cliente/BuscarCliente.java
@@ -1,4 +1,4 @@
-package com.good.food.usecase;
+package com.good.food.usecase.cliente;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -8,13 +8,13 @@ import lombok.RequiredArgsConstructor;
 
 @Component
 @RequiredArgsConstructor
-public class CadastrarCliente {
-
-  @Autowired
-  private final ClienteDatabaseGateway clienteDatabaseGateway; 
+public class BuscarCliente {
   
-  public Cliente execute(Cliente cliente) {
-    return clienteDatabaseGateway.save(cliente);
+  @Autowired
+  private final ClienteDatabaseGateway clienteDatabaseGateway;
+
+  public Cliente findByCpf(final String cpf) {
+    return clienteDatabaseGateway.findByCpf(cpf);
   }
   
 }
